+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Challenge category multiselect formatting for participants
 - Join issues caused by setting a program or branch with an authorization code
 - Secret code field showing when it shouldn't in Mission Control activity logging
+- Single program not being selected when adding household members in Mission Control
 
 ## [4.6.0] 2024-10-09